+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changes:
 
 * It is now possible to configure the JSON Processor used by Pedestal
 * Expanded routes now, again, include the :path-re key (as in Pedestal 0.7 and earlier)
+* The result of `expand-routes`, the RoutingTable record, is now public in the io.pedestal.http.route.types namespace
 
 [Closed Issues](https://github.com/pedestal/pedestal/issues?q=is%3Aclosed%20milestone%3A%220.8.2%22)